+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on Keep a Changelog and this project adheres to Semantic Ver
 ### Added
 - Add Igloo structure retrieval to VanillaStructureProvider (as test case)
 - Add structure preview for those providing structure schematic data
+- Add structure preview window when clicking on the structure preview area
 - Add biome, dimension, rarity information to StructureInfo and display in GUI
 
 ### Fixed